last home

+++ controllers/cards.js
@@ -1,16 +1,17 @@
 const Card = require('../models/card');
+const NotFoundError = require('../errors/not-found-err');
+const AccessDenied = require('../errors/access-denied');
+const SomeError = require('../errors/some-error');
 
-
-const handleResponse = (req, res) => {
+const handleResponse = (req, res, next) => {
   req
     .then((card) => {
-      if(card.length <=0) {
-        res.send({ message: "Карточек нет" })
-      } else {
-        res.send({ data: card })
+      if (!card) {
+        throw new SomeError('Произошла ошибка');
       }
+      res.send({ data: card });
     })
-    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+    .catch(next);
 };
 
 const getCards = (req, res) => {
@@ -18,41 +19,51 @@ const getCards = (req, res) => {
 };
 
 
-const getCard = (req, res) => {
+const getCard = (req, res, next) => {
   const { cardId } = req.params;
-  handleResponse(Card.find({_id: cardId}), res);
+  Card.find({ _id: cardId })
+    .then((card) => {
+      if (card.length <= 0) {
+        throw new NotFoundError('Нет карточки с таким id');
+      } else {
+        res.send({ data: card });
+      }
+    }).catch(next);
 };
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const owner = req.user._id;
+  // if (!req.body.name || !req.body.link) {
+  //   next(new SomeError('Вы не указали название карточки или ссылку на нее'));
+  // } else {}
   const { name, link } = req.body;
   handleResponse(Card.create({ name, link, owner }), res);
 };
 
-const deleteCard = (req, res) => {
+const deleteCard = (req, res, next) => {
   const { cardId } = req.params;
   const currentUser = req.user._id;
-  Card.findOne({ _id: cardId })
-    .then((card) => {
-      const cardOwner = card.owner.toString();
-      //сравниваем ID текущего юзера и создателя карточки
-      if (cardOwner !== currentUser) {
-        //если не совпадает, отправляем сообщение
-        res.status(403).send({ message: 'Отказано в доступе' });
-        //и тут нужно прервать промис
-        return Promise.reject(new Error(''));
-        // reject(new Error('Не вы создавали не вам и удалять'));
-      }
-      //если совпадает, идет then и ищется нужная карточка
-    }).then(() => {
+  Card.findById(cardId).then((card) => {
+    if (!card) {
+      throw new NotFoundError('Нет карточки с таким id');
+    }
+    const cardOwner = card.owner.toString();
+    if (cardOwner !== currentUser) {
+      //  если не совпадает, отправляем сообщение
+      next(new AccessDenied('Отказано в доступе'));
+    } else {
       Card.findByIdAndRemove(cardId)
-        .then((card) => {
-          //показываем карточку
-          res.send({ data: card });
-          //catch при ошибке поиска карточки
-        }).catch((err) => { throw err; });
-      //catch при ошибке в верификации
-    }).catch(() => { res.status(404).send({ message: 'Карточка не найдена' }); });
+        .then((deletedCard) => {
+          if (!deletedCard) {
+            throw new SomeError('Не удалось удалить');
+          }
+          //  показываем карточку
+          res.send({ data: deletedCard });
+          //  catch при ошибке поиска карточки
+        })
+        .catch(next);
+    }
+  }).catch(next);
 };
 
 const likeCard = (req, res) => {
